Cleanup apt install script

+++ installers/apt.sh
@@ -3,6 +3,7 @@
 sudo apt update
 sudo apt upgrade
 
+# Essentials
 sudo apt install -y \
     afl \
     build-essential \
@@ -19,17 +20,12 @@ sudo apt install -y \
     graphviz \
     htop \
     lcov \
-    libboost-all-dev \
-    liblzma-dev \
-    libsqlite3-dev \
-    libxml2-dev \
     lldb \
     lm-sensors \
     nasm \
     nfs-common \
     nodejs \
     python \
-    python-dev \
     python-pip \
     re2c \
     scrot \
@@ -44,38 +40,60 @@ sudo apt install -y \
     xz-utils \
     zsh
 
-# Install latest Vim from source
-# https://github.com/Valloric/YouCompleteMe/wiki/Building-Vim-from-source
-sudo apt remove vim vim-runtime gvim
+# Pass any argument to this script for a "full" install
+if [ -n "$1" ]; then
+    # Developer packages
+    sudo apt install -y \
+        bzip2 \
+        libatk1.0-dev \
+        libbonoboui2-dev \
+        libboost-all-dev \
+        libcairo2-dev \
+        libgnome2-dev \
+        libgnomeui-dev \
+        libgtk2.0-dev \
+        liblzma-dev \
+        libncurses5-dev \
+        libsqlite3-dev \
+        libx11-dev \
+        libxml2-dev \
+        libxpm-dev \
+        libxt-dev \
+        python-dev \
+        ruby-dev \
+        uuid-dev
 
-which vim
-if [ $? != 0 ] ; then
-    sudo apt install -y libncurses5-dev libgnome2-dev libgnomeui-dev \
-        libgtk2.0-dev libatk1.0-dev libbonoboui2-dev \
-        libcairo2-dev libx11-dev libxpm-dev libxt-dev python-dev \
-        ruby-dev
+    # OpenVPN
+    sudo apt install -y \
+        openvpn \
+        network-manager-openvpn \
+        network-manager-openvpn-gnome
 
-    git clone https://github.com/vim/vim.git ~/vim
-    cd ~/vim || exit 1
-    ./configure --with-features=huge \
-                --enable-multibyte \
-                --enable-rubyinterp \
-                --enable-pythoninterp \
-                --with-python-config-dir=/usr/lib/python2.7/config \
-                --enable-perlinterp \
-                --enable-luainterp \
-                --enable-gui=gtk2 --enable-cscope --prefix=/usr
-    make -j VIMRUNTIMEDIR=/usr/share/vim/vim80
-    sudo make install
+    # Vim (from source)
+    # https://github.com/Valloric/YouCompleteMe/wiki/Building-Vim-from-source
+    sudo apt remove vim vim-runtime gvim
+    which vim
+    if [ $? != 0 ] ; then
+        git clone https://github.com/vim/vim.git ~/vim
+        cd ~/vim || exit 1
+        ./configure \
+            --with-features=huge \
+            --enable-multibyte \
+            --enable-rubyinterp \
+            --enable-pythoninterp \
+            --with-python-config-dir=/usr/lib/python2.7/config \
+            --enable-perlinterp \
+            --enable-luainterp \
+            --enable-gui=gtk2 --enable-cscope --prefix=/usr
+        make -j VIMRUNTIMEDIR=/usr/share/vim/vim80
+        sudo make install
 
-    sudo update-alternatives --install /usr/bin/editor editor /usr/bin/vim 1
-    sudo update-alternatives --set editor /usr/bin/vim
-    sudo update-alternatives --install /usr/bin/vi vi /usr/bin/vim 1
-    sudo update-alternatives --set vi /usr/bin/vim
+        sudo update-alternatives --install /usr/bin/editor editor /usr/bin/vim 1
+        sudo update-alternatives --set editor /usr/bin/vim
+        sudo update-alternatives --install /usr/bin/vi vi /usr/bin/vim 1
+        sudo update-alternatives --set vi /usr/bin/vim
+    fi
 fi
-
-# Setup Sensors
-sudo sensors-detect
 
 # Cleanup
 sudo apt autoremove
